Fix Scope
Fix an issue where the ignore configuration didn't support directories.
Directories are only supported on standalone ignore configurations, not in configurations that contain other settings (such as rules or parser options).
Modernize the parser detector so that the TSConfig root dir wasn't necessary as the system can infer it based on node execution context.