Fix Bugs
Fix a bug where the build command didn't work as it didn't have the source code to build from.
Fix another issue here the publish job couldn't complete as it could not find the cache information. Turns out we don't need that so removed the cache requirement as we are only trying to execute npm publish, not npm install.