Optimize Production Build
On the output of the files, the JS Docs were being stripped. This is not ideal as it reduces the intelisense integrations.
Fix the build output so that it no longer nests the binaries in a src folder, and now the prod bins are now in the bin folder, reducing structural complexity on publish.